+++ libraries/data/include/mcrl2/data/detail/prover/info.h
@@ -83,7 +83,7 @@ class Info
           auto index = std::distance(f_variables.begin(), res);
           if (res != f_variables.end())
           {
-            return 1 + (index / ((double)f_variables.size()));
+            return 1 + (static_cast<double>(index) / static_cast<double>(f_variables.size()));
           }
           return 2;
         }
