+++ libraries/pg/include/mcrl2/pg/SmallProgressMeasures.h
@@ -156,7 +156,7 @@ public:
     ParityGame::Player player() const { return (ParityGame::Player)p_; }
 
     /*! Return the length of the SPM vectors (a positive integer). */
-    int len() const { return len_; }
+    int len() const { return static_cast<int>(len_); }
 
     /*! Returns the SPM vector space; an array of len() integers. */
     const verti *M() const { return M_; }
@@ -170,7 +170,7 @@ public:
 
     /*! Return the number of odd priorities less than or equal to the
         priority of v. This is the length of the SPM vector for `v`. */
-    int len(verti v) const { return (game_.priority(v) + 1 + p_)/2; }
+    int len(verti v) const { return static_cast<int>((game_.priority(v) + 1 + p_)/2); }
 
     /*! Return whether the given SPM vector has top value. */
     bool is_top(const verti* vec) const { return vec[0] == NO_VERTEX; }
